requirements.txt
fastapi
uvicorn
qdrant-client
fastembed==0.6.0
python-dotenv
pydantic
sentence-transformers